lib: split out advisories to own file
